backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
gunicorn==20.1.0